requirements.txt
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
coverage==7.13.5
cryptography==46.0.7
docker==7.1.0
Flask==2.3.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
greenlet==3.4.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.1
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==3.0
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-cov==7.1.0
python-dotenv==1.2.2
requests==2.33.1
SQLAlchemy==2.0.49
testcontainers==4.14.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
wrapt==2.1.2
